+++ CHANGELOG.md
@@ -18,6 +18,10 @@
   - [#843](https://github.com/wix-incubator/rich-content/pull/843) fix: mouse up event on overlay triggered the closing the modals
 - `fullscreen`
   - [#848](https://github.com/wix-incubator/rich-content/pull/848) fullscreen is responsive
+### :house: Internal
+- `examples`
+  - [#850](https://github.com/wix-incubator/rich-content/pull/850) fix performance
+  - [#850](https://github.com/wix-incubator/rich-content/pull/850) save content to local storage
 
 </details>
 <hr/>
@@ -78,6 +82,7 @@
   - [#816](https://github.com/wix-incubator/rich-content/pull/816) wrapper theme: gallery settings modal
 - `e2e`
   - [#795](https://github.com/wix-incubator/rich-content/pull/795) new images-sizes fixture
+  - [#759](https://github.com/wix-incubator/rich-content/pull/759) adding tests for emoji plugin (in rtl and plugins tests)
 
 ## 7.0.2 (Mar 20, 2020)
 ### :house: Internal